refactor(web): move bot KPI module from osidb_bot to aegis_ai_web
The KPI computation was incorrectly placed in the osidb_bot package,
which runs in a separate container with its own OSIDB access. Move
all logic into aegis_ai_web where the endpoint lives, and fix test
isolation by patching the cache layer in endpoint tests.
